+++ dataloom/tests/sqlite3/test_qb_sqlite.py
@@ -1,0 +1,58 @@
+class TestQBSQlite:
+    def test_qb(self):
+        from dataloom import (
+            Loom,
+            Model,
+            Column,
+            PrimaryKeyColumn,
+            CreatedAtColumn,
+            TableColumn,
+            ForeignKeyColumn,
+            ColumnValue,
+            UpdatedAtColumn,
+        )
+
+        from typing import Optional
+
+        sqlite_loom = Loom(dialect="sqlite", database="hi.db")
+
+        class User(Model):
+            __tablename__: Optional[TableColumn] = TableColumn(name="users")
+            id = PrimaryKeyColumn(type="int", auto_increment=True)
+            name = Column(type="text", nullable=False, default="Bob")
+            username = Column(type="varchar", unique=True, length=255)
+
+        class Post(Model):
+            __tablename__: Optional[TableColumn] = TableColumn(name="posts")
+            id = PrimaryKeyColumn(
+                type="int", auto_increment=True, nullable=False, unique=True
+            )
+            completed = Column(type="boolean", default=False)
+            title = Column(type="varchar", length=255, nullable=False)
+            # timestamps
+            createdAt = CreatedAtColumn()
+            updatedAt = UpdatedAtColumn()
+            # relations
+            userId = ForeignKeyColumn(
+                User, type="int", required=True, onDelete="CASCADE", onUpdate="CASCADE"
+            )
+
+        conn, _ = sqlite_loom.connect_and_sync([Post, User], drop=True, force=True)
+        userId = sqlite_loom.insert_one(
+            instance=User,
+            values=ColumnValue(name="username", value="@miller"),
+        )
+
+        for title in ["Hey", "Hello", "What are you doing", "Coding"]:
+            sqlite_loom.insert_one(
+                instance=Post,
+                values=[
+                    ColumnValue(name="userId", value=userId),
+                    ColumnValue(name="title", value=title),
+                ],
+            )
+        qb = sqlite_loom.getQueryBuilder()
+        res = qb.run("select id from posts;", fetchall=True)
+        assert str(qb) == "Loom QB<sqlite>"
+        assert len(res) == 4
+        conn.close()